Add FHIR STU3 Examples
Update to latest cql-exec-fhir and add FHIR STU3 versions of each of the examples.
