+++ README.md
@@ -20,9 +20,9 @@ To use this project, you should perform the following steps:
 2. Install [Yarn](https://yarnpkg.com/en/docs/install)
 3. Execute the following from this project's root directory: `yarn`
 
-# Running the Example
+# Running the Examples
 
-The first time you run the examples, you need to download the referenced value sets from the Value Set Authority Center
+The first time you run each example, you need to download the referenced value sets from the Value Set Authority Center
 (VSAC).  This requires your username and password:
 
 ```bash
@@ -33,7 +33,7 @@ NOTE: You can also set the credentials by setting `UMLS_USER_NAME` and `UMLS_PAS
 of passing the credentials on the command line.
 
 After running once with your credentials, the downloaded value set definitions are cached.  After that, you should be
-able to run without credentials:
+able to run the same example without credentials:
 
 ```bash
 $ node index.js
@@ -41,7 +41,9 @@ $ node index.js
 
 # Linting the Code
 
-To encourage quality and consistency within the code base, all code should pass eslint without any warnings.  Many text editors can be configured to automatically flag eslint violations.  We also provide an npm script for running eslint on the project.  To run eslint, execute the following command:
+To encourage quality and consistency within the code base, all code should pass eslint without any warnings.  Many text
+editors can be configured to automatically flag eslint violations.  We also provide an npm script for running eslint on
+the project.  To run eslint, execute the following command:
 
 ```
 $ yarn lint
