+++ diabetic-foot-exam/dstu2.js
@@ -1,0 +1,2 @@
+const runner = require('./runner');
+runner('dstu2');
