+++ diabetic-foot-exam/stu3.js
@@ -1,0 +1,2 @@
+const runner = require('./runner');
+runner('stu3');
