+++ index.js
@@ -1,15 +1,13 @@
 const readline = require('readline');
+const dfeRunner = require('./diabetic-foot-exam/runner');
+const pfRunner = require('./patient-finder/runner');
 
-function printMenu() {
-  console.log('/-------------------------------------------------------------------------------');
-  console.log('| CQL Execution Examples');
-  console.log('\\-------------------------------------------------------------------------------');
-  console.log();
-  console.log('[1] Diabetic Foot Example (FHIR DSTU2)');
-  console.log('[2] Patient Finder');
-  console.log('[Q] Quit');
-  console.log();
-  console.log('Choose an example to run: ');
+function exit(err) {
+  if (err) {
+    process.exit();
+  } else {
+    process.exit(1);
+  }
 }
 
 readline.emitKeypressEvents(process.stdin);
@@ -18,18 +16,35 @@ process.stdin.on('keypress', (c) => {
   console.log();
   switch(c) {
   case 'Q', 'q':
-    process.exit();
+    exit();
     break;
   case '1':
-    require('./diabetic-foot-exam');
+    dfeRunner('dstu2', exit);
     break;
   case '2':
-    require('./patient-finder');
+    dfeRunner('stu3', exit);
+    break;
+  case '3':
+    pfRunner('dstu2', exit);
+    break;
+  case '4':
+    pfRunner('stu3', exit);
     break;
   default:
     console.log('Invalid option: ', c);
+    console.log();
+    exit(1);
   }
-  process.stdin.pause();
 });
 
-printMenu();
+console.log('/-------------------------------------------------------------------------------');
+console.log('| CQL Execution Examples');
+console.log('\\-------------------------------------------------------------------------------');
+console.log();
+console.log('[1] Diabetic Foot Example (FHIR DSTU2)');
+console.log('[2] Diabetic Foot Example (FHIR STU3)');
+console.log('[3] Patient Finder (FHIR DSTU2)');
+console.log('[4] Patient Finder (FHIR STU3)');
+console.log('[Q] Quit');
+console.log();
+console.log('Choose an example to run: ');
