+++ package.json
@@ -12,9 +12,9 @@
     "lint:fix": "./node_modules/.bin/eslint . --fix"
   },
   "dependencies": {
-    "cql-exec-fhir": "^1.0.0",
-    "cql-exec-vsac": "^1.0.0",
-    "cql-execution": "^1.2.1"
+    "cql-exec-fhir": "^1.1.0",
+    "cql-exec-vsac": "^1.0.1",
+    "cql-execution": "^1.3.2"
   },
   "devDependencies": {
     "eslint": "^4.19.1"
